@@ -1,58 +1,45 @@
 {
-  "name": "@blackbox-vision/ra-firebase-starter",
-  "version": "2.0.0",
-  "dependencies": {
-    "@blackbox-vision/ra-language-spanish": "^3.16.2",
-    "@blackbox-vision/ra-layout": "0.0.30",
-    "@blackbox-vision/ra-rbac": "^0.2.0",
-    "@date-io/date-fns": "^1.3.13",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/pickers": "^3.3.10",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^17.0.12",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@vitejs/plugin-react": "^4.2.1",
-    "date-fns": "^2.28.0",
-    "firebase": "^9.6.4",
-    "history": "^4.10.1",
-    "ra-data-json-server": "^3.19.2",
-    "ra-i18n-polyglot": "^3.19.7",
-    "react": "^17.0.2",
-    "react-admin": "^3.19.2",
-    "react-admin-firebase": "^3.2.27",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.5.5",
-    "vite": "^5.2.11",
-    "vite-tsconfig-paths": "^4.3.2",
-    "web-vitals": "^1.0.1"
-  },
+  "name": "ra-firebase-starter",
+  "private": true,
   "scripts": {
-    "start": "vite",
-    "build": "tsc && vite build",
-    "serve": "vite preview"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "type-check": "tsc --noEmit",
+    "lint": "eslint --fix --ext .js,.jsx,.ts,.tsx ./src",
+    "format": "prettier --write ./src"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+  "dependencies": {
+    "@blackbox-vision/ra-language-spanish": "^4.11.3",
+    "@blackbox-vision/ra-rbac": "^4.0.0",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@mui/material": "^5.15.18",
+    "@mui/styles": "^5.15.18",
+    "@mui/x-date-pickers": "^7.5.0",
+    "dayjs": "^1.11.5",
+    "react": "^17.0.0 || ^18.0.0",
+    "react-admin": "^4.16.0",
+    "react-admin-firebase": "^4.1.3",
+    "react-dom": "^17.0.0 || ^18.0.0"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "peerDependencies": {
+    "react": "^17.0.0 || ^18.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "^18.16.1",
+    "@types/react": "^18.0.22",
+    "@types/react-dom": "^18.0.7",
+    "@typescript-eslint/eslint-plugin": "^5.60.1",
+    "@typescript-eslint/parser": "^5.60.1",
+    "@vitejs/plugin-react": "^4.0.1",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^2.8.8",
+    "typescript": "^5.1.6",
+    "vite": "^4.3.9"
   }
 }
