@@ -1,4 +1,4 @@
-import { makeStyles } from "@material-ui/core";
+import { makeStyles } from "@mui/styles";
 
 export let useStyles = makeStyles({
   sidebarWhenClosed: {
