@@ -1,6 +1,6 @@
 import React from "react";
-import { CoreLayoutProps } from "react-admin";
-import { Layout } from "@blackbox-vision/ra-layout";
+import { CoreLayoutProps, Layout } from "react-admin";
+// import { Layout } from "@blackbox-vision/ra-layout";
 
 import { useMenuItemsContext } from "../../../core/context/menu";
 
@@ -10,6 +10,7 @@ type Props = CoreLayoutProps & {
   [key: string]: any;
 };
 
+// TODO: upgrade @blackbox-vision/ra-layout lib dependences to allow use Layout component
 // TODO: check when sidebar is open to pass style information
 let AppLayout: React.FC<Props> = (props: Props) => {
   let menuItems = useMenuItemsContext();
@@ -29,7 +30,5 @@ let AppLayout: React.FC<Props> = (props: Props) => {
     />
   );
 };
-
-AppLayout.displayName = "AppLayout";
 
 export default AppLayout;
