@@ -1,6 +1,9 @@
 import spanishMessages from "@blackbox-vision/ra-language-spanish";
+import users from "../../../../pages/users/locales/es";
 
 export let esMessages = {
   ...spanishMessages,
-  resources: {},
+  resources: {
+    users,
+  },
 };
