@@ -1,15 +1,14 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-import viteTsconfigPaths from "vite-tsconfig-paths";
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
 
+// https://vitejs.dev/config/
 export default defineConfig({
-  // depending on your application, base can also be "/"
-  base: "",
-  plugins: [react(), viteTsconfigPaths()],
-  server: {
-    // this ensures that the browser opens upon server start
-    open: true,
-    // this sets a default port to 3000
-    port: 3000,
-  },
+    plugins: [react()],
+    define: {
+        'process.env': process.env,
+    },
+    server: {
+        host: true,
+    },
+    base: './',
 });
